requirements.txt
python-dotenv
slack-sdk
schedule
